Validate date strings (dates, cancellations)
First check that dates only contain date characters (digits and /).

Then actually try to parse the date.

Note that Date.strptime is quite forgiving and will match a substring
which contains something which looks like a date, ignoring the rest. I
don't want to do something like "parse the date and check that when we
convert it back to a string it's the same", because eg 10/02/2012 and
10/2/2012 would not match, and I think we should be more forgiving than
that.