Disallow REALLY old dates when editing
Ok, so maybe it's not pragmatic to disallow past dates when editing, but
we can prevent mistakes like 01/01/202, which parses as a REALLY old
date.

Anything before the site launched is obviously wrong. I mean, yes, there
are ancient events in there with dates from before launch, and 13 years
ago is a very generous cutoff, but let's see how we go.