Fix failing spec when clicking "Show"
The issue here is that sometimes the event names generated by faker
include the word "Show".

In this case, with the seed 37153, the event generated for the middle
spec for "Archive" was "Dr Low's Medicine Show", and since that was the
first link in the page containing the text "Show", that's the link which
got clicked, and then failed since the URLs aren't real URLs.