Document the need for a NodeJS buildpack
I don't like this.

The issue is that we have one CSS file which is referenced from a
Javascript library, and after moving to DartSASS, that failed on Heroku
build, while passing perfectly fine locally:

    Error: Can't find stylesheet to import.
          ╷
       10 │ @use "accessible-autocomplete/dist/accessible-autocomplete.min.css";
          │ ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

The same applied to @import (necessarily without the .css, to make sure
it's a SASS @import not a plain CSS import:
https://sass-lang.com/documentation/at-rules/import/#importing-css), and
any combination I could think of involving directly referencing the file
in node-modules as a relative path etc.

I believe the reason is that JS isn't actually installed at the point at
which the CSS is built.

jsbundling-rails ensures that `yarn install` is run before
`assets:precompile`.

dartsass-rails is ensures that `dartsass:build` is run before
`assets:precompile`

...but it seems like `dartsass-rails` wins, which means that it's trying
to build CSS partly based on node-modules before js dependencies have
been installed - hence the failure.