Use @use for loading plain css from node-modules
If we use @import we need to omit the .css because otherwise it's
interpreted as a plain CSS import, not a SASS import.

Seems like @use is perfectly fine for importing plain CSS - not sure why
the migrator left this one as an @import:

  https://sass-lang.com/documentation/at-rules/use/#loading-css

Also use @use for the overrides file - that's just a normal SASS file -
not sure why the migrator used meta.load-css.