Refactor: DatesStringValidator delegates to form
Create a form object for validating individual dates.

Note that if we actually try to show `.messages` for these dates then it
will fail since when a symbol is passed, Rails expects it to be a
translation key. But at least at the moment we're never doing that so
maybe it's YAGNI to try.

I guess maybe the desscriptions for the aggregated dates errors would
also live in I18n, but for now it feels simpler in a hash, since that
also gives us a hard exit in case there's an unexpected validation
error, rather than just returning a "translation not found" error.

There's coupling between DatesStringValidator and DateForm: they need to
have exactly the same list of error types. That's fine while DateForm is
a nested class?

Note that we need to move the require of the shared examples before the
require for the validator, so that we can avoid having to include all
the validators individually.