Refactor: Same code used to parse & validate dates
It's a bit sketchy if we validate dates using one approach, and then
later parse them with another: there's a possibility that the parsed
data contains invalid values.

This solution is on the one hand a bit gnarly and not super
straghtforward to get your head around, but on the other hand I'm fairly
pleased with how it feels like it's "Open-Closed": the extra validation
of the date values is added with a block, and the return value is
controlled by an injected object.

Note that we only call uniq in the parse case: it's not clear what the
behaviour should be around duplicates when validating, so I guess in the
scenario where a duplicated date was invalid, we'd report that twice. No
big deal.