Remove concepts of Event inactive? and one_off?
I think the distinction between inactive? and ended? was something which
made sense back when I was proactively looking for events which were
missing dates.

It looks like the intention was something like "find events which aren't
ended, but are only expected to run once, so there's no point looking
for more dates".

...but on closer inspection it looks like the only scenario where any
events would be `inactive?` but not `ended?` are where all dates are empty
(and that's indeed the case on production):

If there are no future dates, and latest_date == last_date then
last_date must be in the past, so by definition the event is ended.
(unless everything is blank).

`one_off?` is more well-defined (apart from matching things with all
blank dates) but just isn't used or useful now.