Extract event updating to a service
Same approach as EventCreator

Needed to change a spec to not try adding a cancellation and date at the
same time. That's a known issue - not sure why it wasn't failing before.

The issue is that we can't have the dates and the cancellations both
trying to create a SwingDate record for the same date.