Remove #dates= and #cancellations= from event
We're now doing this through the creator and updater object, so these
are no longer called except in specs.

The new approach to setting up date data in specs feels awkward, but I'd
argue that it was already awkward, but that awkwardness was hidden by
the fact that these methods were defined on Event.

Note that it turns out that we don't have to blank out the array and add
SwingDate records from scratch: apparently Rails just deals with it.