Handle requests for /.well-known
This is information for robots and API clients - we get a lot of
requests for these files and directories.

These are the most common ones, and those which I know I don't want to
do anything about. I don't want to wildcard this because I think it's
interesting to see what requests are being made.