Add event instances for cancellations on create
Note that for weekly events we're not expecting any dates, so we just
ignore them (that's a problem to be solved earlier - by validation in
the form). Instead we create cancelled event instances for all the given
cancellation dates

For occasional events, we ignore all cancelled dates which don't match
an upcoming date, which is the same as what we were doing before, just
by a more explicit mechanism.

I think this class should assume that any time there's a cancellation
date, that date is always in the upcoming dates, so it shouldn't make
any difference. That's another problem to be solved by validation in the
form.