Don't use `assume_ssl`
The comment is confusing, but this option is only for when the site is
eg. behind a load balancer. As I understand it, the point is that
traffic between the load balancer and the server is over HTTP, but the
original request was via HTTPS, so the server should trust this.

The issue this caused was that force_ssl was never actually redirecting
http to https. This is apparently because the AssumeSSL middleware
pretends that it _already is_ SSL:

  https://github.com/rails/rails/blob/v7.1.2/actionpack/lib/action_dispatch/middleware/assume_ssl.rb#L17-L18