Use event instances
Remove any usage of swing_dates and swing_cancellations.

Note that weekly events can have event instances, but should not have
any dates (we validate that in the form objects), so we return an empty
array for Dates. This will all go away if we eventually create event
instances in the view, rather than two arrays of dates.