Refactor: Occasional means frequency == 0
All new events are weekly (frequency 1) or occasional (frequency 0).

There are still events with a frequency of 2, 4 or more, but they can't
be saved without changing the frequency. Frequency 2 we need to handle
case-by-case, but no events with other frequencies have future dates, so
won't show up on the listings unless new dates are added, which will
force the editor to set the frequency to 0.

I guess previously we've been assuming that any event with dates must be
occasional, and that's hopefully true, but this change makes things more
explicit, and in any case, if an event is weekly and has dates, that
should be considered invalid data.