Validate that cancelled dates are in dates
Dates in the cancelled list will be effectively ignored: cancellations
only have any effect if they match a data in the Upcoming Dates list.

The goal here is to ensure that all cancellations added will actually
get shown.

Note that we need to fix up some specs which broke this new rule.

Note that we need to loosen the assertions of date string validation,
since we will now get extra errors when testing cancellations.

I tried an approach where we only test cancellation date strings for
weekly events, but that doesn't work for the organiser edit form, where
the frequency can't be set in the form.

The "is valid" constraints are now a bit rubbish, but I think not
fragile, since if there are other errors which need to be ignored when
checking date strings, they will result in test failures.

Also putting 'parsed dates' as public attributes is not great, but
otherwise we're duplicating the parsing logic in a _third_ place?
(technically we're parsing as part of validating the string).

Extra nasty is defining day= on the organiser form just to allow us to
use the same validator. But until I see a more elegant way to handle
this, that can stay.