Refactor spec: use frequency traits
Move the event form factories into the main factory and include that:
we can't define a `:weekly` factory in both places because it will get
an error when running the whole suite saying that it's duplicated.

An alternate approach would be to extract a "shared" factory file, but
I think there's value in keeping everything in the same file since it's
so small.