Use event instances for cancelled events
This feels both hacky and very neat: using `select` on a join, exposes
the selected fields as attr readers on the instance, so we can call
Event#cancelled, even though that's not normally a property of the
Event.

The benefit is that we only have to iterate through the event instances
list once, and we get a much cleaner interface for asking whether an
event is cancelled or not on a particular date.

The join on the weekly socials is less neat: we want to do a left join,
since we want to keep all the selected Events regardless of whether they
have any event_instances or not (the instances only exist to indicate if
the event is cancelled on the specific date), and apparently in Rails 7
there's no way to nicely specify conditions _on the join_: if we just
add a `where` then that gets applied to the whole query and ends up
filtering out all weekly events which don't have an associated event
instance.

  https://stackoverflow.com/a/73623452

We need to use `sanitize_sql` because `join` doesn't accept these
interpolation arguments.

We also need to stop using instance doubles, since we're not returning a
normal Event record anymore, so the verifying doubles feature raises an
error.