Block spam/hack attempts with rack-attack
We get a lot of suspicious requests which I can see in Rollbar as
routing errors.

The overwhelming majority of Routing errors we get are wordpress paths
or php files. Fail2ban is kind of neat: you immediately get a 403, and
if you do several within the time period, _all_ requests you make to the
site will be blocked.

Ideally this would be done before the server even gets hit, but this
should significantly reduce the amount of crap we get in our error logs.