Include sitemap in Robots.txt
We get requests form bots for sitemaps at different URLs sometimes.
Let's mitigate that by telling them where to expect to find it.

The challenge is that we need a different URL for London and Bristol, so
to support this, the robots txt needs to be dynamic, but let's ensure
that it's cached and publicly cacheable by intermediate proxies
(whatever those are).