Fix: events with no first date are started
Previously the "if nil?" check was never actually triggered, since we
were checking for the same thing when calling .started? and always
choosing the other option.

This makes sense: if there's no first date then we assume that the first
date is some time in the past.

There's a very slight awkwardness in that when we're _using_ first_date,
the _check_ that we're doing is related to first_date, but that check is
hidden from us inside event.

There's maybe a case for making `started?` private to the presenters,
but it feels like it's a useful question to be able to ask of an event
in general rather than being specific to a particular presentation of
the event.