Set an enum for Event#day
This feels enough: since the user is picking from a select box, not
entering a value, so they can't recover. For that reason it feels
appropriate to get an ArgumentError rather than a validation error.

Note that the change in the Audit logs in specs is because now day will
be saved as nil, while currently it's always stored as an empty string.

Not sure if it's a bad idea to have an enum where the keys and values
are the same, but I can't see a reason why.