Validate first and last date
These always end up as dates or nil in the database, so the risk isn't
of invalid data getting into the database - rather the risk is of
silently ignoring the user's input if it's invalid (since Rails will
just cast it to nil).

Note that we don't do any validation on the Event model itself. This is
a bit tricky: the model casts the string to a date value before any
validation happens.  There's an approach which involves using this:

    record.attributes_before_type_cast[field.to_s]

...and doing the validation on the string before it's cast, but there's
no equivalent of this on ActiveModel, so we couldn't do that on the
forms.