Use date inputs in event form
Turns out that browsers now have pretty good native support for a
datepicker - just adding an input with type "date" is enough.

This results in sending an ISO 8601 formatted date to the form object,
so we need to change up our validation accordingly.

Note that we leave the form objects as handling text fields, rather than
casting them to date types: this lets us validate the values before
they're ultimately type cast (in the model).

This is probably largely academic: we're not expecting to get invalid
dates from the browser (the date field should now catch that, as long as
JS is enabled, which it should be in order for the rest of the form to
work), and if they do, then the validation error is kind of useless,
since there's nothing the user can do about it.

Note that the organiser event form specs weren't using JS before but
they need to now (?) because of the date input.