Use en_GB on CI
This is annoying: the date inputs use a date format based on browser
settings, but setting a `--lang=en_GB` attribute on the selenium chrome
driver doesn't do the trick.

Setting the locale _on the actual CI container_ does, but we need to
install the locale.

I guess ideally we'd make the specs format-agnostic, in case we add a
site or a dev not in the UK, but that's YAGNI and if that does happen
then we've got bigger changes to make.