Move time formats from intializer to I18n
time_formats.rb does still seem to be a recommended approach:

  https://api.rubyonrails.org/classes/Date.html#method-i-to_formatted_s

...but on the other hand I18n.l seems to be the more normal way to
handle date formatting.

The ordinal approach is taken from Rails' own `long_ordinal` format:

  https://github.com/rails/rails/blob/6b93fff8af32ef5e91f4ec3cfffb081d0553faf0/activesupport/lib/active_support/core_ext/date/conversions.rb#L15-L18

Losing the explicit reference to dates as being :uk_date feels in some
way worse, but that is the de-facto default format and a bunch of specs
would need to be updated if we change it later on.

Leave the :db formatting with to_fs because it's language-agnostic.