Show a flash message if event couldn't be archived
This is preferable to the previous behaviour where it would silently do
nothing.

Note that the feature actually wasn't working because it was missing the
turbo: true. A bit worrying that it worked in the specs - I guess
because they're not running with JS, but how is this working _at all_
without js???