Refactor: don't define "today" in controllers
I think at the time I had this idea of passing it in as a dependency
from the very top level, which in a Rails app is the controller. I don't
think this is actually valuable: yes, global state is to be avoided in
general, but time is generally implicitly global state.

Not sure I really have a strong argument, but the code is simpler this
way (according to the ABC metric at least) so I think that justifies
this change.