Rubocop: Address ABC size
With the maps controller, I think the idea before was "pass EVERYTHING
down from the top level (the controller). I no longer believe that
that's valuable.

I think I also got caught up in "Everything is capture/concat". Some
things are just strings.

The event updater thing is odd - not sure if this is a mistake, or if
there was no reason to add the audit comment _after_ setting the
instances before.