Fix Locale issues
In production we have

    config.i18n.fallbacks = true

...which for some reason results in this getting passed as a second
parameter to the lambdas:

    {:fallback_original_locale=>:en}

I guess because maybe we're not explicitly setting locale? Is Rails
automatically setting the locale based on the browser settings? Doesn't
entirely make sense.

Also set the available locales to :en - this is already how it is on
staging (not sure how - maybe heroku does it?), but locally _all_ the
locales seem to be available by default.