Fix external event flash
This was unstyled before, since it was passing a flash key of "success".
It seems like the convention is to only use "alert" and "notice", and to
use "notice" for success.

That seems fine - at least for now: I don't have any general
informational notices.

Chose a more muted colour - a bit less lime.