Copy image for old opengraph image
This is a bit of a hack, particularly since this codebase is used to
serve up both London and Bristol.

The issue:

Facebook expects any OpenGraph image it has seen to be around for ever.

Rails generates cache-busting keys for an image every time it changes.

So we've changed the image several times, but now FaceBook is trying to
fetch the old image, since it thinks of it as a different image.

I don't know if the 74b... image was a version of the current image or
something different, but it's probably the square image we generated
when we didn't know what we were doing, so for now let's just copy the
current opengraph image (which is landscape and nicely compatible with
sharing on Facebook), and expose it under the URL that Fb is expecting
it at.

Another reason this is hacky is that there isn't even supposed to be
anything inside public/assets, but we already have some files there (to
make sure we can render a styled 404 and 500 page when the rest of the
site is broken), so YOLO.