Remove Facebook like button
FB like button doesn't work for me anymore - not sure how long it's been

    broken: Refused to display 'https://www.facebook.com/' in a frame
    because it set 'X-Frame-Options' to 'deny'.

Might be related to the GDPR changes around Like buttons, which to be
honest I don't understand:

  https://developers.facebook.com/docs/plugins/like-button/

Also the number reported on the button was 1.1k, which is incorrect (the
likes on the SODLN FB page is more like 5k) - not sure what to make of
that.

Discussed with Chris and decided it's not valuable anymore, so let's
just cut it.