Fix double factory loading
Sometimes we're getting failures like:

    FactoryBot::DuplicateDefinitionError:
      Factory already registered: event

I think it's to do with random ordering: if the form specs run before
the main system specs run before rails helper is run for the first time,
or the other way round, then we're trying to load the factories when
they've already been loaded, and we get a ton of secondary errors:

    FrozenError:
      can't modify frozen Array:

The idea here is to behave slighly differently depending on whether
we're just running one spec, or if we're doing the whole run. This isn't
theoretically perfect: if we're only running a set of form specs then we
don't need to load the whole Rails environment. However, in practice,
I'm almost always running just one spec file or the whole suite, so this
is maybe good enough.

Previously we had tried to fix this by conditionally only loading
factories in the form spec if they weren't already loaded, but that
didn't actually fix the underlying issue: we still had the scenario
where one of the form specs ran before any of the rails_helper specs
ran.

Most people avoid this issue entirely by always loading the rails helper
in every spec, even the ones which don't need it, and then have Spring
deal with preloading it. But I still think that's slower than it needs
to be and I guess I'm stuck in my ways.