Fix: use new selenium headless chrome
Looks like there was some change in headless chrome where some features
were removed, including execute_cdp, which we rely on for the tests
which exercise the clipboard. There's an old mode which is faster but
doesn't have all the features of a real browser, and a new mode which
does.

  https://github.com/SeleniumHQ/selenium/issues/13112
  https://github.com/SeleniumHQ/selenium/pull/13271

Maybe a better approach would be to keep the old mode for fast tests
which don't need these features and use new mode for tests which do.