Use Capybara from main: handle chromedriver issue
3.40 will use the new headless chrome, but it hasn't been merged yet.

As far as I can gather there was some change in Chrome where they split
headless into two versions: old and new. The old one is used by default,
but is missing some features were removed, including execute_cdp, which
we rely on for the tests which exercise the clipboard This PR explains
it pretty well:

  https://github.com/SeleniumHQ/selenium/pull/13271

What makes this confusing is that the headless browser is provided by
the selenium-webdrivers gem, but driver is specified by Capybara:

  https://github.com/teamcapybara/capybara/commit/e2f9f6cc60d622baf27624e5d224d38cc85f551f

The result was that we got failures on some environments saying that
execute_cdp is not defined - this is presumably due to the Chrome
version being updated - independent of the codebase.

Note that unfortunately the "new" headless browser is slower, but I
think it's probably worth it for keeping the config simple and using a
more realistic browser.