Write some sensible map specs
The previous spec relied on the ancient, pre-capybara method of checking
the values of instance variables. Now we can write some much more
sensible specs for the same functionality, using a custom matcher to
match the page description metadata, since Capybara doesn't provide a
matcher for that (it's mostly concerned with what users can see).

Now we can finally remove that gem which was there to support the
old-style instance variable assertions.