Admins can see the names of users
Note that the cassette has been edited to use dummy IDs and fake names.

The bearer tokens are (presumably?) short-lived, but let's filter them
out anyway.

To record this we needed to actually log in in local development (using
Ngrok) so that we get a real auth token which we can use to make
GraphAPI requests.