Refresh Facebook Graph token
This is hard to test in real life because it seems like Facebook only
issues one token a day: if you request more than one, then the
`expires_in` offset returned results in the same expiry time.

Could go to another level and extract an object which just takes a
login_session and updates it with a new token, but that feels
overengineered for the moment.