Display FB access token expiry in words
This is not entirely accurate: in the 0 case, we know that the token
will expire within 24 hours of now, which could be today or tomorrow.

We could make this more accurate by getting the actual Date objects
associated with these timestamps and comparing those, but I'm not sure
it's worth it? The nice thing about just doing numeric comparisons is
that this object is nice and simple and doesn't need to know about time
zone.