Refactor: move audit urls to presenter
It feels like these should be on the individual Auditable classes, since
we have this repeated case statement, but actually things get a lot more
fiddly if we do that since eg. at the moment we still need to build urls
even for deleted records (the Atom feed uses it as part of the ID, and
always has a url, though looking at the spec, it actually doesn't need
one).

Also we have this fiddliness of ideally wanting to stub a UrlHelpers
instance passed into the AuditLogEntry class, but then needing to pass
that same instance down to the Auditable classes.

Leaving it at the higher level feels ok for now.