Filter HTTP logs
The HTTP gem logs EVERYthing, including bearer tokens.

This is maybe not an actual problem since currently we have log level as
info in production in order to not max out the limits on papertrail. But
I don't want to change that later and unexpectedly leak data.

The solution is to create a log formatter which wraps the normal basic
formatter and sanitises the message.