Client-side validation: reasonable dates
Apply the same validation in the front-end as in the backend

There's a little risk here, since we've been using the start of the Unix
epoch as a way to effectively archive events, and these won't be
updateable anymore. But I guess we shouldn't be touching those unless we
want to resurrect them anyway?