Manage roles in the db rather than ENV
In order to be able to change user permissions, easily remove users, and
eventually improve the onboarding process, we need to start storing the
users in the database rather than just storing IDs in the environment.

A tradeoff is that now a bunch of tests involve an extra database
write/read which they didn't need before, but I guess it's worth it.