Store role in database, not the session
Sessions are long-lived, so we don't want to lock the role into a
session cookie on the user's machine and thus have no way to change it.
Also session manipulation is apparently a thing.'

So instead, any time we need to check the role, we should look in the
database.