Admins can delete users
In order to actually log the user out, we check for the role on every
page load as part of authentication, not just when logging in or when
accessing an admin page.

A user is now not considered logged in unless a role exists in the
database for them.

Actually we were already checking on every page load because of the
navigation: we ask if the user is an admin in order to show the admin
menu options, but this way we can handle things more elegantly.

Notes:
- Needed to add some VCR config - I think in order to use the sessions.
- Define to_param so that we can pass the user object to the URL builder