Refactor: Markdown template handler
I think originally I just copy-pasted this without understanding
anything.

Template handlers are weird: they produce a string which is supposed to
be executed? Let's hide all that nonsense by passing the result to the
Raw handler, which is effectively what the Html handler does in the
Rails source code (albeit with inheritance).