Refactor: extract a Map object
To encapsulate the map setup, and an API object to abstract away the
loading of the map script.

We invert the responsibility for wrangling marker data: previously
createMarkers did the mapping and the creation, but now we separate
those and pass marker data in the correct format to the Map object.
Except for position, since we need to wait until the Google maps API is
available before we can do that.