Float map sidebar over the map
On larger monitors, or just for the classes menu, it doesn't reach the
bottom of the page, so rather than having a big old empty sidebar we can
have the sidebar float over the top of the map.

There are a few things we need to do to compensate:

- the bounding box for markers needs to be offset so that the center of
  the map is where it was before (relative to the sidebar)
- If the sidebar actually _is_ larger than the viewport then we want it
  to scroll, not overflow the page.

Pass the offset as a function, so that the Map can calculate the offset
dynamically whenever it sets bounds padding but doesn't need to be
concerned with how it's calculated.

This messes with the initial center of the map which we set, since we
can't really account for that: it would depend how much of the map is
showing. But the user basically never sees those coordinates anyway:
only if they jump to eg. a day or date on which there are no events, in
which case we don't care so much?