Fix: Map: "All" works correctly with date in URL
The issue seems to be that if there was a date in params, then that gets
passed to the url helper by default, which I guess in other situations
would be what we want, but here we specifically want the url to not
include a date so that we get all dates.

The solution is to split the url helpers so that we explicitly only get
dates when we ask for them.