Service for sending reminder emails to organisers
There's quite a lot here, but it's mostly just setting up the basic
infrastructure required to send mail.

Nothing calls this service yet.

Seems to be possible to just give the gmail account as the mail server?
How does that work? Maybe it's fine as long as it's below a certain
volume of emails?

Previews are at

  http://localhost:3000/rails/mailers/organiser_mailer/reminder_email.html