Override Capybara selenium_chrome_headless driver
The most recent versions of Chrome prompt the user to select a search
engine when opening the browser for the first time. This happens even in
tests, causing tests to fail with errors like "window failed to close
after 20 seconds"

To to disable the chrome search engine selection modal we need to define
a new driver, since this option doesn't seem to be used in the
selenium_chrome_headless driver (which is defined by Capybara).

If Capybara adds the --disable-search-engine-choice-screen option as a
default, we can delete this override.