Verify that doubled constants exist
If a doubled constant doesn't exist then Rails won't verify methods on
it, so we generally want to verify that they do exist. However, we
_don't_ want to verify them in two scenarios:

  - when running isolated spec files (since those constants won't be
    loaded, so would result in verification failures)
  - when TDDing a file (isolated or not) which uses collaborators which
    don't exist yet

In both of those scenarios we're running a single file, so we can use
that as a proxy.

Note that if we were to add it to rails_helper then it would only be
applied to files which were loaded after the first file which loaded
rails helper, which would lead to inconsistent behaviour because of our
random loading.