Validation: ensure no dates past last date
Last date is a source of some confusion. Also it's a hard cut-off: dates
past that date would be ignored.

So instead of this confusing situation where you can add dates which
will never be listed (because they're past the last date), let's make it
so that it's not possible to have dates after the last date.

Note that to make this work we relax the definition of `future_dates?`,
but it will mostly become implicitly still true if we clean up old data.
The scenario where it's not still true is if we have a far future
last_date, but still a few future listing dates before then.

Actually I don't think we need to clean up old data: no current events
violate this constraint and any old dates can deal with it when they get
updated.

One spec was set up which would have violated this new constraint.