Extract find_event to keep ABC metric low
I don't really like this - but it's more like a vague unease. Like it
actually _is_ reasonable to assume that the event ID will be fetched
from params in the same way in all actions of the controller: that's not
just incidental duplication, right?

I like it better than assigning the instance variable in the private
method in any case.