Fix: updates to dates expire the event cache
If there are changes to dates, we want caches involving the event to
expire - in particular in the big admin table, we want the dates to show
up.

I tried quite hard to write a spec for this, but couldn't work out how
to enable the cache. Just stubbing Rails.cache doesn't seem sufficient -
maybe because the fragment caching seems to directly reference the
cache_store config option rather than the currently configured cache?