Fix: generate sensible cache keys
If we pass an array of activerecord objects to the `cache` method in the
view, then Rails produces a cache key by concatenating the values of
`cache_key_with_version` to those objects.

But since we replaced the first item of this array with a presenter
instance, instead the cache key is the serialized array of the full
contents of all the objects, which is crazy and can't even run with the
local in-memory adapter (the browser just shows "try again later").

Defining this method on the presenter allows it to be used as part of
the cache key.

This feels a bit dodgy - I don't think I can write a spec for this cache
key directly, so maybe this could go wrong again and I wouldn't notice.

I could build the cache key in the presenter, but in some sense it feels
like the view should be defining the objects used in the key, since the
view is the one consuming the presenter methods, so is the authority on
what actually needs to be part of it?

...or not, because the view should be dumber than that and shouldn't
know about the relationship between cache objects and methods?

It feels like Rails wants us to put touch: :true on all the associated
relationships, so that any change to a venue would expire the caches for
all associated events. I guess that's what we're doing here too, but
adding those touches would make it a global association and I'm not sure
that's what I want.