Basic email validation
I can see the argument that including a dependency effectively just to
run a regexp is dumb, but it's a problem that I don't want to have to
deal with myself, and feels like a great example of a generic problem
that third party code should deal with.

Also for the create/update forms maybe we'll want to actually validate
that the email address domain actually handles email, and this gem does
that.