Only send one reminder email per week
People will get annoyed if they get an email every day (and Heroku's
free scheduler only allows daily or hourly scheduling).

So let's record when we send emails, and do nothing when trying to send
email before the week is up.

It would be conceptually cleaner if we did the setting of the timestamp
and the checking of the timestamp in the same class.

One option would be to give up on using deliver_later and define an
ActiveJob job - these support after_perform callbacks.

The other would be to move the guard into the mailer.

Neither of these feels good.

Note that for some reason on CI we get a diff in the event spec:

    -Sun, 08 Sep 2024 18:13:32.557836622 BST +01:00
    +Sun, 08 Sep 2024 18:13:32.557836000 BST +01:00

So we use ISO format to ignore the fractional seconds