Validations: handle one-offs
One of the editors has been doing a cludge to add one-off workshops by
creating a weekly class, but then setting the first and last dates to
the same date. This is not OK.

There's also an odd edge case I hadn't considered: listing a social as
weekly but making it a one-off by adding first and last dates the same:
this should have been an occasional event.

This isn't foolproof: if they just remove the start date then the event
will be valid.

Maybe weekly events with an end date in the next week should be invalid,
but that feels like it's going down a path of madness.