Upgrade all gems on 7.1
The only issue is with Rubocop.

I'm generally in favour of sticking to the community defaults, but
you'll pry my String instance doubles out of my cold dead hands: they
allow specs to run completely in isolation.

The enum issue was autofixed and would have caused an exception in a
later rails.