Validate Facebook ID when creating user
We had an issue on production where Chris added "meli.rousseau" as the
Facebook ID. This is obviously incorrect since IDs are numeric strings,
but worse, this record couldn't be deleted - maybe because of how the
dot in the name was interpreted (user couldn't be found).

This can all be avoided by adding some validation.