Don't include the Facebook test user in list
Facebook very annoyingly require us to keep a test user account that
they can log into to verify that the site is legit.

However, for reasons I don't understand, the Graph API calls to fetch
its username fail, leading to errors every time we load the page.

Also someone deleted it by accident since it looks like just junk.

The solution is to special-case it and remove it from the page
completely.