Assert page change before checking record
This failed on mhy new M3 Mac, maybe because of better parallelisation
or just a faster processor? Checking for the page that loads after the
record is created guarantees that we'll have a record to inspect -
otherwise this was turning up as nil.

Also take this opportunity to use `sole`, which is more expressive.