Specs: check for presence of info not absence
Checking for absence is always a bit risky, since content is absent
before the previous page has been replaced. Also there's an inherent
issue in Capybara with the fact that it first finds the content then
makes assertions about it, and sometimes we get errors related to "node"
not being in the page where the page has been switched out in-between
these steps (due to a reload) or part of the page has been switched out
(with JS).

Finally, sometimes we're clicking to view the listings _before the last
event has actually been created_. Actively checking for a "completed"
message avoids that.