Fix User/VCR specs
There are two issues here which can happen intermittently due to fast
hardware executing the spec faster than the app can keep up with:

1. Cassette early exit:

The login step happens inside the cassette block, but we exit the block
before the actual loading of the users happens, since we don't have any
assertion inside the cassette block.

This results in VCR complaining that we're making an API request outside
of a cassette block.

Adding an assertion on content which only exists in the loaded page (the
"Add user" button) means that we only exit the cassette block after the
page has loaded.

2. Checking negative assertion before positive assertion

Whenever we make a change that results in a page load we need to first
make a positive assertion (to ensure that the page has loaded) before we
make any negative assertions ("page doesn't contain...").

Positive assertions wait for the content to appear, but negative
assertions fail as soon as they find that the content is still in the
page (because the change hasn't actually happened yet).